devtools/client/shared/components/reps/test/mochitest/test_reps_object-with-url.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test ObjectWithURL rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - ObjectWithURL</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  try {
    let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    let React = browserRequire("devtools/client/shared/vendor/react");

    const {
      REPS,
      getRep,
    } = browserRequire("devtools/client/shared/components/reps/reps");
    let { ObjectWithURL } = REPS;

    let gripStub = {
      "type": "object",
      "class": "Location",
      "actor": "server1.conn2.obj272",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 15,
      "preview": {
        "kind": "ObjectWithURL",
        "url": "https://www.mozilla.org/en-US/"
      }
    };

    // Test that correct rep is chosen
    is(getRep(gripStub), ObjectWithURL.rep, "Rep correctly selects ObjectWithURL Rep");

    // Test rendering
    const renderedComponent = renderComponent(ObjectWithURL.rep, { object: gripStub });
    ok(renderedComponent.className.includes("objectBox-Location"), "ObjectWithURL rep has expected class name");
    const innerNode = renderedComponent.querySelector(".objectPropValue");
    is(innerNode.textContent, "https://www.mozilla.org/en-US/", "ObjectWithURL rep has expected inner HTML structure and text content");

    // Test rendering with objectLink
    const objectLinkRenderedComponent = renderComponent(ObjectWithURL.rep, {
      object: gripStub,
      objectLink: (props, ...children) => React.DOM.span({},
        "*", ...children, "*"),
    });
    is(objectLinkRenderedComponent.textContent,
      "*Location *https://www.mozilla.org/en-US/",
      "ObjectWithURL rep has expected text content when an objectLink is passed as a prop"
    );

    // @TODO test link once Bug 1245303 has been implemented.
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
